RollingBitFIeld to its own file
