+++ runtime/src/lib.rs
@@ -43,6 +43,7 @@ mod nonce_keyed_account;
 mod pubkey_bins;
 mod read_only_accounts_cache;
 pub mod rent_collector;
+mod rolling_bit_field;
 pub mod secondary_index;
 pub mod serde_snapshot;
 mod shared_buffer_reader;
